pyproject.toml
[project]
name = "cc-feishu-bridge"
version = "0.6.5"
description = "Claude Code Feishu Bridge — 在飞书中与本地 Claude Code 对话"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "claude-agent-sdk>=0.1.59",
    "lark-oapi>=1.0.0",
    "pyyaml>=6.0",
    "python-dateutil>=2.8.0",
    "qrcode>=7.4.0",
    "Pillow>=10.0.0",
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    "filelock>=3.13.0",
    "jieba>=0.42.1",
    "scikit-learn>=1.0",
    "croniter>=2.0",
]
authors = [
    { name = "Hu1J", email = "2948696442@qq.com" },
]

[project.scripts]
cc-feishu-bridge = "cc_feishu_bridge.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.setuptools.packages.find]
where = ["."]
include = ["cc_feishu_bridge*"]
